+++ test/primer/classify_test.rb
@@ -134,6 +134,21 @@ class PrimerClassifyTest < Minitest::Test
     assert_generated_class("color-blue-5",       { color: :blue_5 })
     assert_generated_class("color-gray-9",       { color: :gray_9 })
     assert_generated_class("color-purple-3",     { color: :purple_3 })
+
+    assert_generated_class("color-text-primary",   { color: :text_primary })
+    assert_generated_class("color-text-secondary", { color: :text_secondary })
+    assert_generated_class("color-text-tertiary",  { color: :text_tertiary })
+    assert_generated_class("color-text-link",      { color: :text_link })
+    assert_generated_class("color-text-success",   { color: :text_success })
+    assert_generated_class("color-text-warning",   { color: :text_warning })
+    assert_generated_class("color-text-danger",    { color: :text_danger })
+    assert_generated_class("color-icon-primary",   { color: :icon_primary })
+    assert_generated_class("color-icon-secondary", { color: :icon_secondary })
+    assert_generated_class("color-icon-tertiary",  { color: :icon_tertiary })
+    assert_generated_class("color-icon-info",      { color: :icon_info })
+    assert_generated_class("color-icon-success",   { color: :icon_success })
+    assert_generated_class("color-icon-warning",   { color: :icon_warning })
+    assert_generated_class("color-icon-danger",    { color: :icon_danger })
   end
 
   def test_bg
@@ -299,38 +314,6 @@ class PrimerClassifyTest < Minitest::Test
     assert_generated_class("bg-blue text-center float-left ml-1 ", { classes: "bg-blue text-center float-left ml-1" })
   end
 
-  def test_limits_allocations
-    # Warm up allocations
-    values = {
-      align_items: :center,
-      align_self: :center,
-      bg: :blue,
-      border: :top,
-      box_shadow: true,
-      col: 1,
-      color: :red,
-      flex: 1,
-      float: :left,
-      font_weight: :bold,
-      font_size: 1,
-      height: :fit,
-      justify_content: :flex_start,
-      m: 1,
-      p: 4,
-      position: :relative,
-      text_align: :left,
-      visibility: :hidden,
-      width: :fit,
-      underline: true,
-      vertical_align: true
-    }
-    Primer::Classify.call(**values)
-
-    assert_allocations 87, within: 4 do
-      Primer::Classify.call(**values)
-    end
-  end
-
   private
 
   def assert_generated_class(generated_class_name, input)
@@ -339,17 +322,5 @@ class PrimerClassifyTest < Minitest::Test
 
   def refute_generated_class(input)
     assert_nil(Primer::Classify.call(**input)[:class])
-  end
-
-  def assert_allocations(count, within:)
-    GC.disable
-    total_start = GC.stat[:total_allocated_objects]
-    yield
-    total_end = GC.stat[:total_allocated_objects]
-    GC.enable
-
-    total = total_end - total_start
-
-    assert_in_delta count, total, within, "Expected between #{count - within} and #{count + within} allocations. Got #{total}"
   end
 end